# Conflicts:
#	ci/travis/ci.sh
#	python/ray/tests/test_stress.py
#	python/ray/tests/test_unreconstructable_errors.py
#	rllib/agents/sac/sac_tf_model.py
#	rllib/agents/trainer_template.py
#	rllib/models/modelv2.py

+++ rllib/policy/tf_policy.py
@@ -184,7 +184,7 @@ class TFPolicy(Policy):
         self._stats_fetches = {}
         self._timestep = timestep if timestep is not None else \
             tf1.placeholder_with_default(
-                tf.zeros((), dtype=tf.int64), (), name="timestep")
+                tf.zeros((), dtype=tf.int32), (), name="timestep")
 
         self._optimizer = None
         self._grads_and_vars = None
